Get y Set Memoria
Modularizo los tipos y los accesos a memoria.
Creo el Get y Set, y paso el archivo de direccion logica a fisica al modulo de memoria.
