+++ VirtualMachine.depend
@@ -7,7 +7,16 @@
 	<stdio.h>
 	<stdlib.h>
 
-1757075004 source:d:\facultad\arquitectura\mv\vm---2025\main.c
+1757334216 source:d:\facultad\arquitectura\mv\vm---2025\main.c
 	<stdio.h>
 	<stdlib.h>
+	"acceso_memoria.h"
+
+1757333873 source:d:\facultad\arquitectura\mv\vm---2025\acceso_memoria.c
+	"acceso_memoria.h"
+
+1757334206 d:\facultad\arquitectura\mv\vm---2025\acceso_memoria.h
+	"tipos.h"
+
+1757334206 d:\facultad\arquitectura\mv\vm---2025\tipos.h
