+++ VirtualMachine.layout
@@ -4,7 +4,7 @@
 	<ActiveTarget name="Debug" />
 	<File name="main.c" open="1" top="1" tabpos="1" split="0" active="1" splitpos="0" zoom_1="0" zoom_2="0">
 		<Cursor>
-			<Cursor1 position="1046" topLine="0" />
+			<Cursor1 position="6147" topLine="0" />
 		</Cursor>
 	</File>
 </CodeBlocks_layout_file>
