+++ acceso_memoria.h
@@ -1,0 +1,5 @@
+#include "tipos.h"
+
+void getVM(VM*,int,int*);
+void setVM(VM*,int,int);
+int logica_fisica(VM,int);
